+++ feincms3_meta/utils.py
@@ -1,7 +1,15 @@
 from __future__ import unicode_literals
 
+import warnings
+
 from django.conf import settings
-from django.utils.html import format_html, mark_safe
+from django.utils.html import format_html, html_safe, mark_safe
+
+
+@html_safe
+class MetaTags(dict):
+    def __str__(self):
+        return format_meta_tags(self)
 
 
 def meta_tags(
@@ -34,7 +42,7 @@ def meta_tags(
     ``title``, ``description`` and ``image``. See
     `The Open Graph protocol <http://ogp.me/>`_ for additional details.
     """
-    meta = {'type': 'website'}
+    meta = MetaTags(type='website')
     meta.update(getattr(settings, 'META_TAGS', {}))
 
     for object in reversed(objects):
@@ -96,4 +104,10 @@ def meta_tags_html(*args, **kwargs):
 
     This function has the same signature as ``meta_tags`` above.
     """
-    return format_meta_tags(meta_tags(*args, **kwargs))
+    warnings.warn(
+        'Use meta_tags instead of meta_tags_html; its return value is also'
+        ' directly usable to render all meta tags at once.',
+        DeprecationWarning,
+        stacklevel=1,
+    )
+    return meta_tags(*args, **kwargs)